Add navigation menu support
I ended up changing the classes to match the ones generated by wordpress and inserting the WLZs through the custom item name.

This is not the best solution because it messes up the css/class sructure and it feels dirty overall.

BUT: I was a quick fix (I didn't had to learn how to write a custom walker) and no-so-tech-savvy-users can change the nav AND the corresponding WLZ

Closes #1